requirements.txt
# 鞋子相似度检测系统 - 依赖包
# Shoe Similarity Detection System - Requirements

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# AI模型
transformers>=4.30.0

# 图像处理
pillow>=10.0.0
rembg[cpu]>=2.0.50

# 数据处理
numpy>=1.24.0,<2.4
scikit-learn>=1.3.0

# 配置文件
pyyaml>=6.0

# 向量搜索
faiss-cpu>=1.7.4  # 如果有GPU，可以改为 faiss-gpu

# 工具库
tqdm>=4.65.0

# Web界面
streamlit>=1.28.0

# Web API
flask>=2.3.0
flask-cors>=4.0.0

# 注意：
# 1. rembg[cpu] 用于CPU背景移除，如果有NVIDIA GPU可以改为 rembg[gpu]
#
# 2. 如果有NVIDIA GPU，建议安装GPU版本的PyTorch：
#    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# 3. 如果有NVIDIA GPU，可以安装 faiss-gpu 替代 faiss-cpu：
#    pip install faiss-gpu
#
# 4. 首次运行会自动下载AI模型（约500MB）
#
# 5. 推荐使用Python 3.10或更高版本

